refactor: extract page serving into server/pages.rs
Templates, User-Agent OS detection for the install command, and static
asset serving move out of server/mod.rs into a sibling of binaries.rs.
server/mod.rs is now a thin router plus the broadcast/delete/Socket.IO
handlers, all of which delegate room state to the rooms module.
