+++ src/server/mod.rs
@@ -3,20 +3,20 @@
 //! This module contains the server implementation using axum + socketioxide.
 
 mod binaries;
+mod pages;
 mod rooms;
 
 use axum::{
     body::Body,
     extract::{DefaultBodyLimit, Path, Query, State},
-    http::{header, HeaderMap, Method, Request, StatusCode},
+    http::{header, HeaderMap, StatusCode},
     response::{IntoResponse, Response},
     routing::{delete, get, post},
     Json, Router,
 };
-use crate::protocol;
+use crate::protocol::{self, EncodedMessage};
 use binaries::BinaryDownloadQuery;
 use rooms::{RoomId, Rooms};
-use rust_embed::Embed;
 use socketioxide::{
     extract::{Data, SocketRef, State as SioState},
     SocketIo,
@@ -27,16 +27,6 @@ use std::time::Duration;
 use tokio::sync::RwLock;
 use tower_http::trace::TraceLayer;
 use tracing::{info, warn};
-
-/// Embedded static files from the public directory
-#[derive(Embed, Clone)]
-#[folder = "public/"]
-struct StaticAssets;
-
-/// Embedded view templates
-#[derive(Embed, Clone)]
-#[folder = "templates/"]
-struct Templates;
 
 /// Room cleanup configuration
 #[derive(Clone, Debug)]
@@ -114,14 +104,14 @@ pub async fn run(
     // Build router
     let app = Router::new()
         // API routes
-        .route("/", get(index_handler))
-        .route("/r/{*room}", get(room_page_handler))
+        .route("/", get(pages::index_handler))
+        .route("/r/{*room}", get(pages::room_page_handler))
         .route("/r/{*room}", post(broadcast_handler))
         .route("/r/{*room}", delete(delete_room_handler))
         // Binary download (serves embedded binaries or self)
         .route("/bin/shellshare", get(serve_binary))
         // Static files - fallback
-        .fallback(serve_static)
+        .fallback(pages::serve_static)
         // State and middleware
         .with_state(app_state)
         .layer(sio_layer)
@@ -232,74 +222,6 @@ fn setup_socket_handlers(io: &SocketIo) {
     });
 }
 
-/// Detect OS from User-Agent header
-fn detect_os_from_user_agent(user_agent: &str) -> &'static str {
-    let ua = user_agent.to_lowercase();
-    if ua.contains("windows") {
-        "windows"
-    } else if ua.contains("mac") {
-        "macos"
-    } else {
-        "linux" // default
-    }
-}
-
-/// GET / - Home page
-async fn index_handler(headers: HeaderMap) -> impl IntoResponse {
-    match Templates::get("index.html") {
-        Some(content) => {
-            let user_agent = headers
-                .get(header::USER_AGENT)
-                .and_then(|v| v.to_str().ok())
-                .unwrap_or("");
-
-            let os = detect_os_from_user_agent(user_agent);
-
-            let html = String::from_utf8_lossy(&content.data);
-            let html = html
-                .replace(
-                    "{{LINUX_CHECKED}}",
-                    if os == "linux" { " checked" } else { "" },
-                )
-                .replace(
-                    "{{MACOS_CHECKED}}",
-                    if os == "macos" { " checked" } else { "" },
-                )
-                .replace(
-                    "{{WINDOWS_CHECKED}}",
-                    if os == "windows" { " checked" } else { "" },
-                );
-
-            Response::builder()
-                .status(StatusCode::OK)
-                .header(header::CONTENT_TYPE, "text/html; charset=utf-8")
-                .body(Body::from(html))
-                .unwrap()
-        }
-        None => Response::builder()
-            .status(StatusCode::INTERNAL_SERVER_ERROR)
-            .header(header::CONTENT_TYPE, "text/plain; charset=utf-8")
-            .body(Body::from("Template not found"))
-            .unwrap(),
-    }
-}
-
-/// GET /r/:room - Room page
-async fn room_page_handler(Path(_room): Path<String>) -> impl IntoResponse {
-    match Templates::get("room.html") {
-        Some(content) => Response::builder()
-            .status(StatusCode::OK)
-            .header(header::CONTENT_TYPE, "text/html; charset=utf-8")
-            .body(Body::from(content.data.into_owned()))
-            .unwrap(),
-        None => Response::builder()
-            .status(StatusCode::INTERNAL_SERVER_ERROR)
-            .header(header::CONTENT_TYPE, "text/plain; charset=utf-8")
-            .body(Body::from("Template not found"))
-            .unwrap(),
-    }
-}
-
 /// POST /r/:room - Broadcast message to room
 async fn broadcast_handler(
     Path(room_path): Path<String>,
@@ -326,7 +248,7 @@ async fn broadcast_handler(
     let message = body_obj
         .and_then(|o| o.get("message"))
         .and_then(|m| m.as_str())
-        .map(|m| crate::protocol::EncodedMessage::from_wire(m.to_string()));
+        .map(|m| EncodedMessage::from_wire(m.to_string()));
 
     // Claim/verify the room and store - atomically
     if state
@@ -425,38 +347,4 @@ async fn serve_binary(
     headers: HeaderMap,
 ) -> impl IntoResponse {
     binaries::serve_binary(query, headers).await
-}
-
-/// Serve embedded static files
-async fn serve_static(req: Request<Body>) -> impl IntoResponse {
-    let path = req.uri().path().trim_start_matches('/');
-    let method = req.method();
-
-    match StaticAssets::get(path) {
-        Some(content) => {
-            // Only allow GET and HEAD for existing static files
-            if method != Method::GET && method != Method::HEAD {
-                return Response::builder()
-                    .status(StatusCode::METHOD_NOT_ALLOWED)
-                    .header(header::ALLOW, "GET, HEAD")
-                    .header(header::CONTENT_TYPE, "text/plain; charset=utf-8")
-                    .body(Body::from("Method Not Allowed"))
-                    .unwrap();
-            }
-            let mime = mime_guess::from_path(path).first_or_octet_stream();
-            let etag = format!("\"{}\"", hex::encode(content.metadata.sha256_hash()));
-            Response::builder()
-                .status(StatusCode::OK)
-                .header(header::CONTENT_TYPE, mime.as_ref())
-                .header(header::CACHE_CONTROL, "public, no-cache")
-                .header(header::ETAG, etag)
-                .body(Body::from(content.data.into_owned()))
-                .unwrap()
-        }
-        None => Response::builder()
-            .status(StatusCode::NOT_FOUND)
-            .header(header::CONTENT_TYPE, "text/plain; charset=utf-8")
-            .body(Body::from("Not Found"))
-            .unwrap(),
-    }
 }
